refactor FlowEditor components
