+++ src/pages/FlowEditor/StepForm/Media.tsx
@@ -1,0 +1,5 @@
+export default function MediaStepForm() {
+  return (
+    <p className="text-sm text-muted-foreground">Formulário de mídia ainda não implementado.</p>
+  );
+}
