+++ src/pages/FlowEditor/StepForm/Text.tsx
@@ -1,0 +1,6 @@
+import type { Step } from "../../../types/flow";
+
+export default function TextStepForm({ step }: { step: Step }) {
+  // Texto não possui campos adicionais além dos básicos
+  return null;
+}
