+++ src/pages/FlowEditor/index.tsx
@@ -2,23 +2,14 @@ import type React from "react";
 
 import { useEffect, useMemo, useCallback, useState } from "react";
 import { useParams, useNavigate, Link } from "react-router-dom";
-import { DndContext, closestCenter, type DragEndEvent } from "@dnd-kit/core";
-import {
-  arrayMove,
-  SortableContext,
-  verticalListSortingStrategy,
-  useSortable,
-} from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
-import {
-  GripVertical,
-  Plus,
-  Trash2,
+import type { DragEndEvent } from "@dnd-kit/core";
+import { Trash2,
   ArrowLeft,
   Menu,
   Eye,
   Settings,
 } from "lucide-react";
+import { arrayMove } from "@dnd-kit/sortable";
 import { nanoid } from "nanoid";
 import { useFlows } from "../hooks/useFlows";
 import { Button } from "../components/ui/button";
@@ -26,7 +17,6 @@ import { Card, CardContent } from "../components/ui/card";
 import { Sheet, SheetContent, SheetTrigger } from "../components/ui/sheet";
 import { Separator } from "../components/ui/separator";
 import { Badge } from "../components/ui/badge";
-import { Skeleton } from "../components/ui/skeleton";
 import {
   AlertDialog,
   AlertDialogAction,
@@ -40,7 +30,10 @@ import {
 } from "../components/ui/alert-dialog";
 import { useToast } from "../hooks/use-toast";
 import { cn } from "../utils/cn";
-import StepForm from "../components/flow/StepForm";
+import StepForm from "./StepForm";
+import StepsSidebar from "./StepsSidebar";
+import EmptyState from "./EmptyState";
+import FlowEditorSkeleton from "./Skeleton";
 import type { Flow, Step } from "../types/flow";
 
 const STEP_TYPES = {
@@ -216,6 +209,7 @@ export default function FlowEditor() {
           onDragEnd={handleDragEnd}
           onStepDelete={handleStepDelete}
           stepIds={stepIds}
+          stepTypes={STEP_TYPES}
         />
       </aside>
 
@@ -255,6 +249,7 @@ export default function FlowEditor() {
                   onDragEnd={handleDragEnd}
                   onStepDelete={handleStepDelete}
                   stepIds={stepIds}
+                  stepTypes={STEP_TYPES}
                   isMobile
                 />
               </SheetContent>
@@ -366,218 +361,3 @@ export default function FlowEditor() {
   );
 }
 
-interface StepsSidebarProps {
-  flow: Flow;
-  selectedId: string | null;
-  onStepSelect: (id: string) => void;
-  onAddStep: () => void;
-  onDragEnd: (event: DragEndEvent) => void;
-  onStepDelete: (id: string) => void;
-  stepIds: string[];
-  isMobile?: boolean;
-}
-
-function StepsSidebar({
-  flow,
-  selectedId,
-  onStepSelect,
-  onAddStep,
-  onDragEnd,
-  onStepDelete,
-  stepIds,
-  isMobile = false,
-}: StepsSidebarProps) {
-  return (
-    <div className="flex flex-col h-full">
-      <div className="p-4 border-b">
-        {isMobile && (
-          <div className="mb-4">
-            <h2 className="font-semibold">{flow.title}</h2>
-            <p className="text-sm text-muted-foreground">
-              {flow.steps.length} {flow.steps.length === 1 ? "passo" : "passos"}
-            </p>
-          </div>
-        )}
-
-        <div className="flex items-center justify-between mb-4">
-          <h3 className="font-medium">Passos do Fluxo</h3>
-          <Button size="sm" onClick={onAddStep}>
-            <Plus className="h-4 w-4 mr-1" />
-            Adicionar
-          </Button>
-        </div>
-      </div>
-
-      <div className="flex-1 overflow-y-auto p-4">
-        {flow.steps.length === 0 ? (
-          <div className="text-center py-8">
-            <p className="text-sm text-muted-foreground mb-3">
-              Nenhum passo criado ainda
-            </p>
-            <Button size="sm" variant="outline" onClick={onAddStep}>
-              <Plus className="h-4 w-4 mr-1" />
-              Criar primeiro passo
-            </Button>
-          </div>
-        ) : (
-          <DndContext collisionDetection={closestCenter} onDragEnd={onDragEnd}>
-            <SortableContext
-              items={stepIds}
-              strategy={verticalListSortingStrategy}
-            >
-              <div className="space-y-2">
-                {flow.steps.map((step: Step, index: number) => (
-                  <StepItem
-                    key={step.id}
-                    id={step.id}
-                    step={step}
-                    index={index}
-                    selected={step.id === selectedId}
-                    onSelect={() => onStepSelect(step.id)}
-                    onDelete={() => onStepDelete(step.id)}
-                  />
-                ))}
-              </div>
-            </SortableContext>
-          </DndContext>
-        )}
-      </div>
-    </div>
-  );
-}
-
-interface StepItemProps {
-  id: string;
-  step: Step;
-  index: number;
-  selected: boolean;
-  onSelect: () => void;
-  onDelete: () => void;
-}
-
-function StepItem({ id, step, index, selected, onSelect, onDelete }: StepItemProps) {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-    opacity: isDragging ? 0.5 : undefined,
-  } as React.CSSProperties;
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className={cn(
-        "group relative rounded-lg border p-3 cursor-pointer transition-all",
-        "hover:border-primary/50 hover:shadow-sm",
-        selected && "border-primary bg-primary/5 shadow-sm",
-        isDragging && "shadow-lg"
-      )}
-      onClick={onSelect}
-    >
-      <div className="flex items-start gap-3">
-        <div
-          className="mt-1 cursor-grab active:cursor-grabbing opacity-40 group-hover:opacity-100 transition-opacity"
-          {...attributes}
-          {...listeners}
-        >
-          <GripVertical className="h-4 w-4" />
-        </div>
-
-        <div className="flex-1 min-w-0">
-          <div className="flex items-center gap-2 mb-1">
-            <span className="text-xs font-medium text-muted-foreground">
-              #{index + 1}
-            </span>
-            <Badge
-              variant="secondary"
-              className={cn(
-                "text-xs px-1.5 py-0.5",
-                STEP_TYPES[step.type as keyof typeof STEP_TYPES]?.color
-              )}
-            >
-              {STEP_TYPES[step.type as keyof typeof STEP_TYPES]?.label}
-            </Badge>
-          </div>
-          <h4 className="font-medium text-sm truncate">{step.title}</h4>
-          {step.content && (
-            <p className="text-xs text-muted-foreground mt-1 line-clamp-2">
-              {step.content}
-            </p>
-          )}
-        </div>
-      </div>
-      <Button
-        size="sm"
-        variant="ghost"
-        onClick={(e) => {
-          e.stopPropagation();
-          onDelete();
-        }}
-        className="absolute top-2 right-2 text-muted-foreground hover:text-destructive opacity-0 group-hover:opacity-100"
-      >
-        <Trash2 className="h-4 w-4" />
-      </Button>
-    </div>
-  );
-}
-
-function EmptyState({ onAddStep }: { onAddStep: () => void }) {
-  return (
-    <Card className="h-full flex items-center justify-center">
-      <CardContent className="text-center py-12">
-        <div className="mx-auto w-12 h-12 bg-muted rounded-full flex items-center justify-center mb-4">
-          <Plus className="h-6 w-6 text-muted-foreground" />
-        </div>
-        <h3 className="text-lg font-semibold mb-2">Nenhum passo selecionado</h3>
-        <p className="text-muted-foreground mb-6 max-w-sm">
-          Selecione um passo na barra lateral para editá-lo, ou crie um novo
-          passo para começar.
-        </p>
-        <Button onClick={onAddStep}>
-          <Plus className="mr-2 h-4 w-4" />
-          Criar primeiro passo
-        </Button>
-      </CardContent>
-    </Card>
-  );
-}
-
-function FlowEditorSkeleton() {
-  return (
-    <div className="flex h-screen">
-      <aside className="hidden lg:flex w-80 border-r p-4">
-        <div className="w-full">
-          <div className="flex items-center justify-between mb-4">
-            <Skeleton className="h-5 w-20" />
-            <Skeleton className="h-8 w-20" />
-          </div>
-          <div className="space-y-3">
-            {Array.from({ length: 3 }).map((_, i) => (
-              <Skeleton key={i} className="h-16 w-full" />
-            ))}
-          </div>
-        </div>
-      </aside>
-      <main className="flex-1">
-        <div className="p-6 border-b">
-          <div className="flex items-center justify-between">
-            <Skeleton className="h-8 w-48" />
-            <Skeleton className="h-8 w-24" />
-          </div>
-        </div>
-        <div className="p-6">
-          <Skeleton className="h-96 w-full" />
-        </div>
-      </main>
-    </div>
-  );
-}
